+++ hyde/model.py
@@ -81,6 +81,12 @@ class Expando(object):
                 result[k] = v
         return result
 
+    def get(self, key, default=None):
+        """
+        Dict like get helper method
+        """
+        return self.__dict__.get(key, default)
+
 
 class Context(object):
     """
@@ -142,6 +148,7 @@ class Config(Expando):
     def __init__(self, sitepath, config_file=None, config_dict=None):
         self.default_config = dict(
             mode='production',
+            simple_copy = [],
             content_root='content',
             deploy_root='deploy',
             media_root='media',