+++ hyde/tests/ext/test_flattener.py
@@ -23,7 +23,7 @@ class TestFlattner(object):
     def tearDown(self):
         TEST_SITE.delete()
 
-    def test_can_flattener(self):
+    def test_can_flatten(self):
         s = Site(TEST_SITE)
         cfg = """
         plugins:
@@ -42,4 +42,26 @@ class TestFlattner(object):
 
         assert not s.config.deploy_root_path.child_folder('blog').exists
         assert File(s.config.deploy_root_path.child('merry-christmas.html')).exists
+
+    def test_flattener_fixes_nodes(self):
+        s = Site(TEST_SITE)
+        cfg = """
+        plugins:
+            - hyde.ext.plugins.folders.FlattenerPlugin
+        flattener:
+            items:
+                -
+                    source: blog
+                    target: ''
+        """
+        import yaml
+        s.config = Config(TEST_SITE, config_dict=yaml.load(cfg))
+        s.load()
+        gen = Generator(s)
+        gen.generate_all()
+        blog_node = s.content.node_from_relative_path('blog')
+        
+        assert blog_node
+        assert blog_node.url == '/'
+